Specify data type in queries for constant values.
git-svn-id: file:///Users/gauthierm/svnrepo/admin/trunk@4248 94490950-a86c-4493-9ccb-530f72933e47

+++ Admin/AdminSummaryDependency.php
@@ -135,7 +135,7 @@ class AdminSummaryDependency extends AdminDependency
 		}
 
 		$sql = sprintf('select count(%s) as count, %s as parent,
-				%s as status_level from %s',
+				%s::integer as status_level from %s',
 			$id_field->name, $parent_value, $status_level, $table);
 
 		if ($where_clause !== null)
